Move persist logic into DB module
