+++ src/db/db.ts
@@ -1,0 +1,7 @@
+import knex from "knex";
+import knexfile from "./knexfile";
+
+const env = process.env.NODE_ENV ?? "development";
+const db = knex(knexfile[env]);
+
+export default db;
